+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-03-11T09:45:14.163Z")
+    updated: Optional[datetime] = parse_datetime("2026-03-11T12:40:16.356Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")