+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKE4JZ72AV27CB4Q928X4GGK"
+        "01KKEEKG1QG3FECPGW9H6AM3VH"
     )
     r"""The unique identifier of the data source"""