+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKE4JZ78CCDQ1J38BTS5DJ71"
+        "01KKEEKG1VNZCTJ5BAN7KJTM8B"
     )
     r"""The unique identifier of the data source"""