+++ packages/orq-rc/RELEASES.md
@@ -6479,3 +6479,13 @@ Based on:
 - [python v4.6.0-rc.1] packages/orq-rc
 ### Releases
 - [PyPI v4.6.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.6.0-rc.1 - packages/orq-rc
+
+## 2026-03-11 12:43:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.751.0 (2.858.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.6.0-rc.2] packages/orq-rc
+### Releases
+- [PyPI v4.6.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.6.0-rc.2 - packages/orq-rc